Changed the getName and getTitle methods to name and title properties. The name property is just the prefix, a dot and the field name. The field name is used to generate request variables, although, widgets using multiple input elements will often need to generate more specific names.
Added haveData method to test whether there is user input.

Added required widget attribute that simply delegates to the
underlying field required attribute.

Added support for CustomWidgets in automatic widget setup code.

Added fieldNames utility function to get the names of fields defined
by a schema in definition order.

Added haveWidgets data to easily tests whether any widgets have
data. This can be used to determine if a form was submitted or is
being viewed for the first time. It's better to use an explicit
request variable, set by a submit button to test this.

Moved the setPrefix, name, and setData machinery to the generic widget
code.


git-svn-id: svn://svn.zope.org/repos/main@6729 62d5b8a3-27da-0310-9561-8e5933582275