Changed to use (single-line) Line field for content type.
Added an automatically-generated upload form that is customized with a
class to provide a custom widget (a file widget) for the data and to
provide size information.  The form is also custimized with a custom
template that adds size information. It uses macros from the base
template.

Had to provide a menu definition for I18nImage to avoid getting menu
definitions from IFile.


git-svn-id: svn://svn.zope.org/repos/main@6735 62d5b8a3-27da-0310-9561-8e5933582275