Fix subtle bug in restore().
The _txn_find() must not stop at the pack boundary when it is called
by restore().  It was originally written for _txn_undo() which isn't
supposed to undo to a transaction across a pack.  But it should be
legal to restore() a transaction with a reference to a data record in
a transaction that was packed.

Fix by adding stop_at_pack flag to _txn_find() and add tests of this
behavior for FileStorage.


git-svn-id: svn://svn.zope.org/repos/main@2489 62d5b8a3-27da-0310-9561-8e5933582275