Removed Christian Theune's XXX comment.
If you comment like this in the source, please post a question to
zope3-dev also, so that someone will answer the question.
It is only by chance that I saw this commit message, and thus could
easily address the answer to the correct person.

Views and Resources and other Presentation components are always concerned
with Requests.

Other things are also concerned with Requests.
The ErrorReportingService is concerned with a Request because just about
everything Zope does is initiated by a Request of some sort.
So, when you log an error, it makes sense to describe the Request that
could not be satisfied due to that error.

There are other services that are concerned with requests. For example,
the Views service.

In general, your content components will not be directly concerned with
requests. It is up to presentation components to extract pertinent
information from the request, and pass just that application-pertinent
information to the content component. So, content components aren't
concerned with requests.


git-svn-id: svn://svn.zope.org/repos/main@7179 62d5b8a3-27da-0310-9561-8e5933582275