A *very* preliminary version of code to create and configure Events and ObjectHub services.  This currently hardcodes everything.  It doesn't activate or even register them, because I don't know how to do that yet.
XXX My motivation for checking this in now is:

(1) It works (as far as it goes).

(2) It's a definite improvement (cuts out several steps from my
    recipe for setting up a text index).

(3) Maybe SteveA or Jim can look at my code and suggest a better way
    to do it, or explain why my attempts to activate the services
    failed.

(4) I plan to generalize this and add ZCML directives, once I
    understand how it's done.

XXX Note: Wednesday I'm traveling to ZC HQ, which means I may not get
to hack on this code until Thursday.


git-svn-id: svn://svn.zope.org/repos/main@7199 62d5b8a3-27da-0310-9561-8e5933582275