Since both the Windows and Unix ZEO tests share the same code, it makes little sense to call the server script `winserver.py'.  That's gone now, in favor of this zeoserver.py script.  Changes include:
main(): Accept a -C option pointing to a temp file containing a
ZConfig storage section.  This reads the file and creates the so
described storage, then removes the temp config file.  It creates an
administrative server which opens an `admin' port which is used to
os._exit() this process, with the following protocol.  The first time
the admin port is connected to, an ack character is sent from
zeoserver.  This is necessary to coordinate the tests so that they are
assured the server is running and responding before it tries to do
more.  Otherwise, some tests were vulnerable to timing bugs where the
shutdown connect happened before the server was ready to accept them.

The second connect exits the server process just like before.  Copious
log entries are written.

This script also cleans up after the storages are closed, by calling
the storage's .cleanup() method if it exists (FileStorage and the
Berkeley storages both implement this method).


git-svn-id: svn://svn.zope.org/repos/main@2505 62d5b8a3-27da-0310-9561-8e5933582275