Further to Jeremy's comment on the nonsensical code here:
http://lists.zope.org/pipermail/zope3-checkins/2002-December/002575.html

Yes, _contextWrapperChecker(ob) should be returning a checker.

defineChecker(_Wrapper, _contextWrapperChecker) won't work, becuase
_Wrapper is a factory function, not a type. This is lameness in the
Zope/ContextWrapper package -- Wrapper was originally both a type and
a factory. Now it is only a factory. You need to iterate through
wrapperTypes to get all of the available types of ContextWrapper.
The interactions of different kinds of proxies and wrappers in Zope 3
needs a fresh look and some refactoring. Let's hope we can get to this
before the alpha.

defineChecker should really check that it is given a type.


git-svn-id: svn://svn.zope.org/repos/main@7293 62d5b8a3-27da-0310-9561-8e5933582275