Apparently functional version that reads from the end.
Starts near the end of the file and reads forward, looking for a
complete snapshot.  If it doesn't find one, move farther from the end
and repeat.

XXX This approach does N**2 reads, where N is the number of times we
need to move further back and retry.  When we move backwards, we
double the distance from the end of the file.


git-svn-id: svn://svn.zope.org/repos/main@2536 62d5b8a3-27da-0310-9561-8e5933582275