SteveA and Marius G.
We refactored the NamedComponentConfiguration and ComponentConfiguration.
Now, we have NamedConfiguration and NamedComponentConfiguration.

NamedConfigurable is now split into two parts, as appropriate to that
refactoring.

Cleaned up various extraneous imports, and changed various
self --> wrapped_self. (Interestingly, this act revealed one ContextMethod
that did not in fact need to be a ContextMethod.)


git-svn-id: svn://svn.zope.org/repos/main@7317 62d5b8a3-27da-0310-9561-8e5933582275