Cleaned up merge of Introspector work in Bangalore branch.  will not be hooked up until the OFS configure.zcml is also committed: I'm going to test my commit before I do so.
When this is hooked up, it will add an Introspector action in the UI.

This html was broken in the Rotterdam skin; fixed and significantly cleaned up.

A lot of DOS line endings were removed.  dos2unix reports that all is well now.

Commented-out code in the view refers to an IConfigureFor interface that goes in a direction that Jim eventually wants.  I am not merging in the necessary code elsewhere (in the services) that this code needs, so therefore I comment it out but do not delete it.

Introspector tab will be available to anonymous user, as it was in the branch.  If we don't like this, we can just fix the zcml.


git-svn-id: svn://svn.zope.org/repos/main@7323 62d5b8a3-27da-0310-9561-8e5933582275