Add a framework for converting asyncore errors to unittest errors.
By default, asyncore handles uncaught exceptions in dispatchers by
printing a message to the console.  If a test causes such uncaught
exceptions, the test is marked as a failure, because asyncore handles
the exception.  This framework causes the unit test to fail.  If code
being tested expects the errors to occur, it can add code to prevent
the error from propagating all the way back to asyncore.


git-svn-id: svn://svn.zope.org/repos/main@7429 62d5b8a3-27da-0310-9561-8e5933582275