Implemented VFS views for the Local Service Manager. You can now traverse into it up to Packages.
Since Jim did not tell me what he wanted to do in a Package, I did not
add much there, so that all Services are shown as directories now. The
only binding I implemented was for the ZPTTemplate, which works fine so
that you can add ZPTs via VFS now. I could imagine doing the same for
Module, but other than that Package is rather boring from a VFS point of
view.

Of course we could do some "magic" and display a text (file content)
representation of the various services, but I do not know how useful this
would be.


git-svn-id: svn://svn.zope.org/repos/main@7467 62d5b8a3-27da-0310-9561-8e5933582275