timetz comparison, and datetimetz subtraction:  reimplemented these to work as documented.  They weren't calling utcoffset() if both operands had the same tzinfo object.  That's fine if it so happens that the shared tzinfo object returns a fixed offset (independent of operand), but can give wrong results if that's not so, and the latter obtains in a tzinfo subclass instance trying to model both standard and daylight times.
This repairs some of the surprises in the sandbox US.py demo, but
not all of them.  In a way, it makes one part even more surprising:
because addition of datetimetz and timedelta doesn't do anything with
tzinfo except attach it to the result, it's possible to add, e.g.,
one second to a datetimetz, and then see the result of that minus the
original datetimetz span an hour.  This can happen at DST boundaries
(see US.py's doctest for concrete examples).


git-svn-id: svn://svn.zope.org/repos/main@8537 62d5b8a3-27da-0310-9561-8e5933582275