Changed the way the icon_url is used.  This variable is computed on the server as an absolute url and it should be used as such.
TODO:

It would be good to compute the expansion icon (pl.gif and mi.gif)
urls in a similar fashion. As it is, they are assumed to be in the
root folder. This means that they can't be changed locally. This is
probably not a big deal.  They could be set by the template page as
the base url and local urls are.


git-svn-id: svn://svn.zope.org/repos/main@8608 62d5b8a3-27da-0310-9561-8e5933582275