The failure of the last-second addition to the timezone coversion test is understood now:  it can't work.  Added comments explaining why (it's "the usual"-- unrepresentable hours in local time --but in a slightly different guise).
Added an optimization to astimezone().  This will pay off more in the C
implementation (which is already optimized to work with offsets as C
ints internally, instead of with timedelta objects; this will let it
replace a datetimetz comparison call with a couple of int operations; in
the Python implementation it trades away a datetimetz comparison call
for a couple of timedelta operations).


git-svn-id: svn://svn.zope.org/repos/main@8738 62d5b8a3-27da-0310-9561-8e5933582275