Refactor ZEOStorage to remove the strategy classes.
The strategy classes stopped being useful, because the code had been
modified to always use one of the strategies.  The classes were left
behind because editing the code to use a single class was too much
work at the time.

Move all the logic back into ZEOStorage.  Rename non-public methods on
the class to start with _.  Use lambda: rather than getattr()() to
invoke the method that gets delayed waiting for the storage lock.


git-svn-id: svn://svn.zope.org/repos/main@2581 62d5b8a3-27da-0310-9561-8e5933582275