Merge from 2_6 branch: fixed stored acquisition wrappers.
Some indexes were storing acquisition wrappers in the cyclic references
back to the ZCatalog instance.  Removed the acquisition wrappers and
cleaned up a bad use of apply().  The acquisition wrappers were revealed
by AdaptableStorage.

There's a second problem, in that indexes are unknowingly storing references
back to the ZCatalog in an unused attribute.  In the 2_6 branch it's
hard to fix without breaking things, but on the HEAD there's a better way
to fix this.  Coming soon, I hope.


git-svn-id: svn://svn.zope.org/repos/main@23407 62d5b8a3-27da-0310-9561-8e5933582275