Get ZERO from the package.
Remove __version__

ZODB4 will have two new methods in the storage api:

    get_version() -> string
    set_version(string)

Prepare for this now by changing the packtime table into an info
table, which is just a key/value mapping for meta-information about
the storage.  For now, the only key we define is `packtime' although
we're reserving `version' for ZODB4.


git-svn-id: svn://svn.zope.org/repos/main@23411 62d5b8a3-27da-0310-9561-8e5933582275