zope.app.form.utility had a function, fieldNames, that returned a list of the Fields in a schema in schema order.  This is exactly parallel to getFieldsInOrder from the schema package itself, and it seems to me that's where it belongs.  So I added a getFieldNamesInOrder function to schema.  I also added a getFieldNames in parallel to getFields, to complete the set.  Everything that used fieldNames is converted to use getFieldNamesInOrder.
The unit tests from form was actually a little stronger than those
in schema:  it made sure that all fields were returned when using
a derived schema, which the schema tests didn't.  So I added some
subschema tests to the schema test suite in addition to the tests
for the new functions themselves.


git-svn-id: svn://svn.zope.org/repos/main@8965 62d5b8a3-27da-0310-9561-8e5933582275