Local event service now uses adapts to ISubscribingAware.
I changed loops of the form

  for subs in something:
      dosomething(subs[0])
      dosomethingelse(subs[1])

to

  for athing, anotherthing in something:
      dosomething(athing)
      dosomethingelse(anotherthing)

However, I don't understand what the 'anotherthing' should be in a couple
of places. I'll look into it further, or ask Gary for help.

Also, there are a couple of places where I'm not sure if the subscriber
object needs to be checked to see if it is None. I need to look into
those cases.

All unit tests pass, however.


git-svn-id: svn://svn.zope.org/repos/main@8974 62d5b8a3-27da-0310-9561-8e5933582275