__getstate__() isn't needed for pickling anymore, but is used for other things (like hashing and comparison), so make them private methods.  This was done earlier for __setstate__() too.
Note that the use of getstate() to get state strings for comparisons (and
such) makes good *sense* only in the 2.3 C implementation, where the
internal object memory *is* a little state string.  We're doing this in
the Python implementation too primarily to make sure it always delivers
the same results as the C implementation.


git-svn-id: svn://svn.zope.org/repos/main@8991 62d5b8a3-27da-0310-9561-8e5933582275