Remove unneeded try/finally.
There is no need to explicitly zap a traceback object passed as an
argument to a function.  In general, if an exception is raised in a
function and that traceback is stored, then the function's frame will
be kept alive because the traceback has a reference to it.  This
function doesn't create a traceback, so it isn't directly responsible
for extending the life of the frame.

It may be the case that the caller of handleException() must be
careful.  On the other hand, the garbage collector will be able to
collect a cycle involving traceback and frame unless some objects with
finalizers are involved.


git-svn-id: svn://svn.zope.org/repos/main@9014 62d5b8a3-27da-0310-9561-8e5933582275