Attempt to reduce the disk I/O caused when running this script, by saving state in a pck pickle file, and doing an incremental parse the next time.  Specifically:
Added -f and -r flags.  The -f flag points to the pickle state file
(by default ./zeoqueue.pck).  The -r flag simply removes this file.
This is useful for log rotation in the cron job, and yes rm would have
been fine, but this option was requested by customers.

process_file(): Added to Status class so that the processing is done
here, where we can seek to a stored file position.

Also added the secret -0 option for testing, which suppresses the file
position seek (useful when the log file is split into chunks).

Return the number of blocked clients as the exit status code.


git-svn-id: svn://svn.zope.org/repos/main@2771 62d5b8a3-27da-0310-9561-8e5933582275