Restructure the default configuration manager view.
This now has two lines per configuration object.  Line one is the
"usage summary", typically showing the kind of configuration and the
key.  Line two is the "implementation summary", showing some of the
additionally configured information.  What exactly is shown is up to
the configuration object.

Adding the methods to come up with the summary lines to the
configuration object rather than making them views is perhaps a little
odd, but Jim decided that it was better than invoking the complicated
machinery needed to define views to return this information.

The "ItemEdit" views are gone, as are some associated templates.

I renamed some of the views to index.html, because we seem to like
that as the default view nowadays.


git-svn-id: svn://svn.zope.org/repos/main@9234 62d5b8a3-27da-0310-9561-8e5933582275