Fix several ref count bugs.
Remove object_from_oid() which appeared to be mis-used about half the
time.  Instead use std PyDict_GetItem() and incref only when
necessary.

Make sure an object's reference to its cache is decref'd when the
object is deallocated.  This change seems to eliminate leaking pickle
caches.

Remaining mystery: Why do objects stay in the cache even when there
are no references to them?


git-svn-id: svn://svn.zope.org/repos/main@2833 62d5b8a3-27da-0310-9561-8e5933582275