Restore cache behavior for unreferenced by recently used objects.
The implementation is to make the reference in the cache "ring" a new
reference rather than a borrowed reference.  It is the intent of the
cache to keep N recently used objects in memory, regardless of whether
they are currently referenced.  The goal is to avoid the cost of
recreating the objects, based on the assumption that they are likely
to be used again soon.


git-svn-id: svn://svn.zope.org/repos/main@2846 62d5b8a3-27da-0310-9561-8e5933582275