Don't unghostify an object for __del__.
Persistent objects are often in cycles -- they point to their data
manager and their data manager points to them.  Python 2.2 will look
for an __del__ attribute on an instance that is unreachable.  We don't
want to unghostify an object in this case.

XXX Should enforce the policy that Persistent objects can't have an
__del__.

XXX Must backport to Zope2.


git-svn-id: svn://svn.zope.org/repos/main@9499 62d5b8a3-27da-0310-9561-8e5933582275