Slightly cleaner implementation of MessageID.  Logically I think these should be immutable (but this is not enforced yet since I'm not sure). However, the domain and default attributes might as well be directly accessible so get rid of the setDefault() method.
Also add an __new__() so that the domain and default can be specified
in the constructor.  The additional benefit here is that domain and
default are always defined, even if they are set to None (the old way,
you'd have to do a hasattr() first).

Updated docstrings.

Update MessageIDFactory.__call__() to use the new signature for
MessageID.

Whitespace normalization.


git-svn-id: svn://svn.zope.org/repos/main@9576 62d5b8a3-27da-0310-9561-8e5933582275