Speed up connection establishment.
Don't know what happened, but the tests started running very slowly
with Python2.3.  Fixed by checking _ready at the start of the loop in
_wait() instead of at the bottom.  Most of the time _ready is set by
the time we start, and no more I/O is going to occur until the client
connects; as a result, the tests always sat in pending() until it
timed out at 30 seconds.


git-svn-id: svn://svn.zope.org/repos/main@2861 62d5b8a3-27da-0310-9561-8e5933582275