The hubIds service depends on the event service. So, add the hubIds service after the event service. Make the event channel (on which the event service depends) not require a hubIds service.
So, most people's hubids services are not subscribed to their event
services. The easy way to fix this is to toss your Data.fs.
The fairly easy way to fix this is to go into the debugger, and subscribe
the service in the course of another transaction.
The pretty easy way to fix this is to make the hubids service disabled,
and then re-enabled.


git-svn-id: svn://svn.zope.org/repos/main@9629 62d5b8a3-27da-0310-9561-8e5933582275