Updated the form-genmeration software to support accessors. Mainly this involved changing the software to use field ``get``, ``query``, and ``set`` methods rather than attribute access.  Also modified the tests to test accessor usage.
Fixed adapter support. Now edit views store the adapted context
separately from the context. Also, edit and add views use getApter
rather than queryAdapter to get adapters. The previous uses of
queryAdapter were hiding some logic errors.  Added unit tests!
In the course of adding unit tests, I discovered an issue relating to
event generation. Normally, we generate an ObjectModification event
when we edit something, however, it's not so clear what we should do
in the presense of an adapter. I finally decided that we should not
generate an event if the object was adapted. In that case, the adapter
should generate any necessary events.


git-svn-id: svn://svn.zope.org/repos/main@9633 62d5b8a3-27da-0310-9561-8e5933582275