Fix a rather subtle bug in ViewService.getRegisteredMatching().
The viewName argument was reused as a loop control variable,
overwriting the search key when searching the second and later
layers.  This caused the View service not to show views defined in
other layers than the default (at least in my case).

XXX This is missing a test case; I don't understand the test code well
enough to be able to add one.


git-svn-id: svn://svn.zope.org/repos/main@9835 62d5b8a3-27da-0310-9561-8e5933582275