Rewrote the support for running doctest tests as unittests.
Now, there is a function for computing a unittest test suite based on
the doctest tests found in a module.

Added a debugger for doctest tests.  This is needed because the
standard Python debugger won't work with doctest, because doctest
hijacks stdout.

Also added a text extractor that extracts a test script from a doctest
test doc string.


git-svn-id: svn://svn.zope.org/repos/main@10080 62d5b8a3-27da-0310-9561-8e5933582275