Moved the iterator code here. Modified iterators to work with Python iterators and iterables. Added doctest tests.
Changed the strategy used to manage variables. Variables
(context.vars) is now just a dictionary, rather than a special
object.  Likewise for the repeat variables.

Added a context hook so that subclasses can override the way context
variables are set. (Specifically so that Zope contexts can put
security proxies around the context (top-level) variables.

Attempted to make the module usable even if zope.tal can't be
imported.

Renamed a bunch of interfaces to reflect the TAL expression, rather
than TALES.

Moved the SimpleExpr class, which is only used for testing to the test
package.


git-svn-id: svn://svn.zope.org/repos/main@10111 62d5b8a3-27da-0310-9561-8e5933582275