Renamed the getAnnotation and hasAnnotation methods of principal annotation services to getAnnotations and hasAnnotations.
The implementation assumed it was getting principal ids even though
the interface said principals were passed and non testing clients
passed principals. Fixed the implementation to expect principals for
getAnnotations and hasAnnotations.

Added a getAnnotationsById, which is needed to do service delegation
in certain situations.

Now new annotations aren't stored unless they are modified. This is to
prevent database writes on "read" requests.


git-svn-id: svn://svn.zope.org/repos/main@10257 62d5b8a3-27da-0310-9561-8e5933582275